Created a Dockerfile.
Created a Dockerfile for the program. Unfortunately, I'm having an issue where the java program cannot find the class within the Docker container. I am very new to Docker and will keep trying to troubleshoot, but I'm taking a break. I believe I am not actually using the right file path in the mv commands in the dockerfile.